config: Expand \~ in zephyr_base from user args
When twister_config is created the valuses from config are taken
"as they are". This caused problems when \~ was is in zephyr_base.
It was not expanded and certain modules using zephyr_base from config
are failing, e.g. cmake_filter. This patch fixes the issue by expanding
the path before it is added to twister_config

Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>